requirements.txt
Flask
psycopg2-binary

Flask==2.3.3
gunicorn==21.2.0
